Fix type issue in pausepi config file

+++ config/pausepi.php
@@ -3,6 +3,6 @@
 declare(strict_types=1);
 
 return [
-    'min_timeout_seconds' => env('PAUSEPI_MIN_TIMEOUT_SECONDS', 30),
-    'max_timeout_seconds' => env('PAUSEPI_MAX_TIMEOUT_SECONDS', 300),
+    'min_timeout_seconds' => (int) env('PAUSEPI_MIN_TIMEOUT_SECONDS', 30),
+    'max_timeout_seconds' => (int) env('PAUSEPI_MAX_TIMEOUT_SECONDS', 300),
 ];
